Move Synth off of the stack.
This is a fairly major change that moves the Synth structure into Rust-managed memory, allowing deeper changes to be made, moving closer towards the goal of making all of the code safe.

As a knock-on effect, the ReverbModel struct had to temporarily be moved to unsafe direct heap allocation. It should be rewritten to use vec probably, which will give better flexibility for parameterization anyways.